# scanner/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: КВАНТ Скан API
  version: "1.0.0"
  description: |
    Контракт между мобильными клиентами, web-консолью, внешними системами
    и бэкендом платформы «КВАНТ Скан».

    Разделы:
      * `sync`     — синхронизация устройств (pull изменений, push событий)
      * `assets`   — чанковая загрузка медиа с докачкой
      * `tasks`    — задания на сканирование
      * `protocols`— протоколы съёмки
      * `items`    — карточки единиц товара и трассировка
      * `inbound`  — приём заданий из складских систем
      * `review`   — контроль качества

    Принципы, зафиксированные в контракте:
      * идемпотентность по `client_event_id` (UUIDv7) и `Idempotency-Key`;
      * результат по каждому событию отдельно — батч не падает целиком;
      * неизвестные поля клиенты обязаны игнорировать (аддитивная эволюция).
servers:
  - url: https://scan.kvant.ru/v1
    description: Промышленная среда
  - url: https://scan-stage.kvant.ru/v1
    description: Тестовая среда
security:
  - bearerAuth: []
...
paths:
  /sync/pull:
    post:
      tags: [sync]
      summary: Получить изменения с момента курсора
      description: |
        Возвращает дельту по заданиям, протоколам и справочникам. Курсор
        непрозрачен для клиента. Сервер использует `capabilities`, чтобы не
        выдавать заданий, которые устройство не сможет выполнить.
      operationId: syncPull
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SyncPullRequest' }
      responses:
        '200':
          description: Дельта изменений
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SyncPullResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /sync/events:
    post:
      tags: [sync]
      summary: Отправить батч событий сессии
      description: |
        Идемпотентно по `client_event_id`. Ответ содержит результат по каждому
        событию отдельно: `accepted`, `duplicate` (тоже успех) или `rejected`.
        HTTP 200 возвращается и при наличии отвергнутых событий — транспорт
        отработал, семантика в теле.
      operationId: syncPushEvents
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EventBatchRequest' }
      responses:
        '200':
          description: Результат по каждому событию
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventBatchResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413':
          description: Батч превышает допустимый размер (по умолчанию 1000 событий / 4 МБ)
        '429': { $ref: '#/components/responses/RateLimited' }
        '503':
          description: Backpressure — очередь переполнена, повторить после `Retry-After`
          headers:
            Retry-After: { schema: { type: integer }, description: Секунды до повтора }
...
  /assets:
    post:
      tags: [assets]
      summary: Инициировать загрузку медиа
      description: |
        Возвращает `asset_id` и параметры чанковой загрузки. Если актив с такой
        `sha256` уже принят в этом тенанте, возвращается `deduplicated: true`
        и загрузка не требуется.
      operationId: initAsset
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssetInitRequest' }
      responses:
        '201':
          description: Загрузка инициирована
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssetInitResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /assets/{assetId}/chunks/{n}:
    put:
      tags: [assets]
      summary: Загрузить чанк
      description: Идемпотентно по паре `(assetId, n)`. Повторная загрузка того же чанка безопасна.
      operationId: putAssetChunk
      parameters:
        - name: assetId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: n
          in: path
          required: true
          schema: { type: integer, minimum: 0 }
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        '204': { description: Чанк принят }
        '409': { description: Размер чанка не совпадает с согласованным }
  /assets/{assetId}/complete:
    post:
      tags: [assets]
      summary: Завершить загрузку и сверить контрольную сумму
      operationId: completeAsset
      parameters:
        - name: assetId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Результат сверки
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [verified, checksum_mismatch, missing_chunks]
                  missing_chunks:
                    type: array
                    items: { type: integer }
  /assets/{assetId}/status:
    get:
      tags: [assets]
      summary: Состояние загрузки (для докачки)
      operationId: getAssetStatus
      parameters:
        - name: assetId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Полученные чанки
          content:
            application/json:
              schema:
                type: object
                required: [asset_id, upload_state, received_chunks]
                properties:
                  asset_id: { type: string, format: uuid }
                  upload_state:
                    type: string
                    enum: [pending, uploading, complete, verified]
                  received_chunks:
                    type: array
                    items: { type: integer }
  /tasks:
    get:
      tags: [tasks]
      summary: Список заданий
      operationId: listTasks
      parameters:
        - name: state
          in: query
          schema: { $ref: '#/components/schemas/TaskState' }
        - name: site_id
          in: query
          schema: { type: string, format: uuid }
        - name: assignee_user_id
          in: query
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 500 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Страница заданий
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ScanTask' }
                  next_cursor: { type: string, nullable: true }
  /tasks/{taskId}:
    get:
      tags: [tasks]
      summary: Задание по идентификатору
      operationId: getTask
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Задание
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScanTask' }
        '404': { $ref: '#/components/responses/NotFound' }
  /protocols/{code}/{version}:
    get:
      tags: [protocols]
      summary: Версия протокола съёмки
      description: |
        Версии иммутабельны и доступны вечно, включая `deprecated` — иначе
        невозможно воспроизвести, по каким инструкциям снимали год назад.
      operationId: getProtocol
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string, pattern: '^[a-z][a-z0-9_]{2,63}$' }
        - name: version
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Документ протокола (см. capture-protocol.schema.json)
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '404': { $ref: '#/components/responses/NotFound' }
  /directories/{name}:
    get:
      tags: [sync]
      summary: Справочник, синхронизируемый на устройство
      description: |
        Таблицы стандартов (ряды ГОСТ 6636, сечения колец, толщины листа и
        т.п.) ведёт владелец реестра как данные — код приложения при их
        расширении не меняется (Р-03 разд. 6). Устройство запрашивает
        справочник на каждой синхронизации с `If-None-Match` и в обычном
        случае получает `304` без тела: версии меняются редко.
      operationId: getDirectory
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string, enum: [standard_series] }
        - name: If-None-Match
          in: header
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Документ справочника; ETag считается из его версии
          headers:
            ETag:
              schema: { type: string }
              description: '`"<имя>-v<версия>"`'
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '304':
          description: Версия на устройстве актуальна, тело не передаётся
        '404': { $ref: '#/components/responses/NotFound' }
  /review/queue:
    get:
      tags: [review]
      summary: Открытые сессии на ручной проверке
      description: |
        Сессии сюда ставит приёмка вердиктом `MANUAL_REVIEW`. Доступно только
        роли `reviewer`; сотрудник заказчика очередь не видит.
      operationId: listReviewQueue
      responses:
        '200':
          description: Очередь, старые сверху
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ReviewEntry' }
        '403':
          description: Роль не reviewer
  /review/{sessionId}/verdict:
    post:
      tags: [review]
      summary: Вердикт контролёра по сессии
      description: |
        `ACCEPTED` — принято; `REJECTED` — забраковано; `REWORK` — вернуть
        на пересъёмку (повторная постановка в очередь переоткрывает запись).
        Повторный вердикт по уже решённой записи — `409`: решение первого
        контролёра не перезаписывается молча.
      operationId: postReviewVerdict
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [verdict]
              properties:
                verdict: { type: string, enum: [ACCEPTED, REJECTED, REWORK] }
                comment: { type: string }
      responses:
        '200':
          description: Запись с вынесенным вердиктом
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewEntry' }
        '403':
          description: Роль не reviewer
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Вердикт уже вынесен другим контролёром
...
  /items/{itemId}/trace:
    post:
      tags: [items]
      summary: Добавить связь трассировки
      description: |
        Доступно в том числе внешним пользователям — но только с
        `confidence: declared`. Значение `verified` может проставить лишь
        роль `reviewer` или машинная сверка с реестром.
      operationId: addTraceLink
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TraceLinkInput' }
      responses:
        '201':
          description: Связь добавлена
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TraceLink' }
        '403':
          description: Попытка выставить уровень доверия, недоступный роли
  /inbound/tasks:
    post:
      tags: [inbound]
      summary: Принять задание из внешней складской системы
      description: |
        Заголовок `Idempotency-Key` обязателен. Повторная доставка того же
        вебхука возвращает тот же `task_id` и не создаёт дубль. Ключ хранится
        30 дней.
      operationId: inboundTask
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, maxLength: 200 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InboundTaskRequest' }
      responses:
        '202':
          description: Задание принято (ещё не назначено)
          content:
            application/json:
              schema:
                type: object
                required: [task_id, state]
                properties:
                  task_id: { type: string, format: uuid }
                  state: { $ref: '#/components/schemas/TaskState' }
                  deduplicated:
                    type: boolean
                    description: true — задание уже существовало, создан не был
        '400':
          description: Ошибка валидации с указанием поля
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Токен отсутствует, истёк или отозван
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    RateLimited:
      description: Превышен лимит запросов
      headers:
        Retry-After:
          schema: { type: integer }
          description: Секунды до повтора
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    SyncPullRequest:
      type: object
      required: [device_id, capabilities]
      properties:
        cursor: { type: string, nullable: true }
        device_id: { type: string, format: uuid }
        capabilities: { $ref: '#/components/schemas/DeviceCapabilities' }
        scopes:
          type: array
          items: { type: string, enum: [tasks, protocols, directories, verdicts] }
    SyncPullResponse:
      type: object
      required: [changes, next_cursor, has_more, server_time]
      properties:
        changes:
          type: array
          items:
            type: object
            required: [type, data]
            properties:
              type:
                type: string
                enum: [task.upsert, task.revoke, protocol.upsert, directory.patch, verdict.upsert]
              data: { type: object, additionalProperties: true }
        next_cursor: { type: string }
        has_more: { type: boolean }
        server_time: { type: string, format: date-time }
...
    EventBatchRequest:
      type: object
      required: [device_id, events]
      properties:
        device_id: { type: string, format: uuid }
        events:
          type: array
          minItems: 1
          maxItems: 1000
          items: { $ref: '#/components/schemas/ClientEvent' }
...
    TaskState:
      type: string
      enum: [NEW, ASSIGNED, IN_PROGRESS, SUBMITTED, IN_REVIEW, REWORK, ACCEPTED, REJECTED, EXPORTED, CANCELLED, EXPIRED]
...